move algorithm utils to separate file
