+++ backend/src/algo/consts.ts
@@ -1,0 +1,2 @@
+export const daysPerWeek = 5;
+export const maxPeriodsPerDay = 8;
